+++ crates/common/src/raindex_client/orders.rs
@@ -57,7 +57,7 @@ pub struct RaindexOrder {
 #[cfg(target_family = "wasm")]
 #[wasm_bindgen]
 impl RaindexOrder {
-    #[wasm_bindgen(getter)]
+    #[wasm_bindgen(getter = chainId)]
     pub fn chain_id(&self) -> u64 {
         self.chain_id
     }
@@ -367,9 +367,8 @@ impl RaindexClient {
         page: Option<u16>,
     ) -> Result<Vec<RaindexOrder>, RaindexError> {
         let raindex_client = Arc::new(RwLock::new(self.clone()));
-        let multi_subgraph_args = self.get_multi_subgraph_args(
-            chain_ids.map(|ids| ids.0.iter().map(|id| *id as u64).collect()),
-        )?;
+        let multi_subgraph_args = self
+            .get_multi_subgraph_args(chain_ids.map(|ids| ids.0.iter().map(|id| *id).collect()))?;
         let client =
             MultiOrderbookSubgraphClient::new(multi_subgraph_args.values().cloned().collect());
         let orders = client
@@ -409,10 +408,10 @@ impl RaindexClient {
 
     async fn get_sg_order(
         &self,
-        chain_id: u16,
+        chain_id: u64,
         order_hash: String,
     ) -> Result<SgOrder, RaindexError> {
-        let subgraph_url = self.get_subgraph_url_for_chain(chain_id as u64)?;
+        let subgraph_url = self.get_subgraph_url_for_chain(chain_id)?;
         let client = OrderbookSubgraphClient::new(subgraph_url);
         let order = client.order_detail_by_hash(SgBytes(order_hash)).await?;
         Ok(order)
@@ -453,13 +452,13 @@ impl RaindexClient {
     )]
     pub async fn get_order_by_hash(
         &self,
-        chain_id: u16,
+        chain_id: u64,
         order_hash: String,
     ) -> Result<RaindexOrder, RaindexError> {
         let raindex_client = Arc::new(RwLock::new(self.clone()));
         let order = RaindexOrder::try_from_sg_order(
             raindex_client,
-            chain_id as u64,
+            chain_id,
             self.get_sg_order(chain_id, order_hash).await?,
             None,
         )?;
